requirements.txt
streamlit
pandas
python-docx
unidecode
openpyxl
